+++ client.diff/src/main/java/com/github/gumtreediff/client/diff/web/VanillaDiffHtmlBuilder.java
@@ -31,9 +31,12 @@ import gnu.trove.map.hash.TObjectIntHashMap;
 import java.io.*;
 import java.nio.charset.Charset;
 import java.nio.file.Files;
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
-public final class HtmlDiffs {
+public final class VanillaDiffHtmlBuilder {
 
     private static final String SRC_MV_SPAN = "<span class=\"%s\" id=\"move-src-%d\" data-title=\"%s\">";
     private static final String DST_MV_SPAN = "<span class=\"%s\" id=\"move-dst-%d\" data-title=\"%s\">";
@@ -52,7 +55,7 @@ public final class HtmlDiffs {
 
     private Diff diff;
 
-    public HtmlDiffs(File fSrc, File fDst, Diff diff) {
+    public VanillaDiffHtmlBuilder(File fSrc, File fDst, Diff diff) {
         this.fSrc = fSrc;
         this.fDst = fDst;
         this.diff = diff;
@@ -164,5 +167,49 @@ public final class HtmlDiffs {
         else if (cr == '>') w.append("&gt;");
         else if (cr == '&') w.append("&amp;");
         else w.append(cr);
+    }
+
+    private static class TagIndex {
+
+        private Map<Integer, List<String>> startTags;
+
+        private Map<Integer, List<String>> endTags;
+
+        public TagIndex() {
+            startTags = new HashMap<Integer, List<String>>();
+            endTags = new HashMap<Integer, List<String>>();
+        }
+
+        public void addTags(int pos, String startTag, int endPos, String endTag) {
+            addStartTag(pos, startTag);
+            addEndTag(endPos, endTag);
+        }
+
+        public void addStartTag(int pos, String tag) {
+            if (!startTags.containsKey(pos)) startTags.put(pos, new ArrayList<String>());
+            startTags.get(pos).add(tag);
+        }
+
+        public void addEndTag(int pos, String tag) {
+            if (!endTags.containsKey(pos)) endTags.put(pos, new ArrayList<String>());
+            endTags.get(pos).add(tag);
+        }
+
+        public String getEndTags(int pos) {
+            if (!endTags.containsKey(pos)) return "";
+            StringBuilder b = new StringBuilder();
+            for (String s: endTags.get(pos)) b.append(s);
+            return b.toString();
+        }
+
+        public String getStartTags(int pos) {
+            if (!startTags.containsKey(pos))
+                return "";
+            StringBuilder b = new StringBuilder();
+            for (String s: startTags.get(pos))
+                b.append(s);
+            return b.toString();
+        }
+
     }
 }